Finish SpecimenRequestManager
