+++ study/api-src/org/labkey/api/specimen/SpecimenRequestException.java
@@ -1,0 +1,13 @@
+package org.labkey.api.specimen;
+
+public class SpecimenRequestException extends RuntimeException
+{
+    public SpecimenRequestException()
+    {
+    }
+
+    public SpecimenRequestException(String message)
+    {
+        super(message);
+    }
+}
