+++ study/src/org/labkey/study/model/StudyImpl.java
@@ -24,7 +24,6 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.junit.Assert;
 import org.junit.Test;
-import org.labkey.api.annotations.Migrate;
 import org.labkey.api.attachments.Attachment;
 import org.labkey.api.attachments.AttachmentFile;
 import org.labkey.api.attachments.AttachmentParent;
@@ -52,14 +51,10 @@ import org.labkey.api.security.SecurityPolicy;
 import org.labkey.api.security.SecurityPolicyManager;
 import org.labkey.api.security.User;
 import org.labkey.api.security.permissions.AdminPermission;
-import org.labkey.api.specimen.SpecimenRequestManager;
-import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
-import org.labkey.api.specimen.settings.RepositorySettings;
-import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.study.Location;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
@@ -78,7 +73,6 @@ import org.labkey.api.util.TestContext;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.wiki.WikiRendererType;
 import org.labkey.api.wiki.WikiRenderingService;
-import org.labkey.study.SpecimenManager;
 import org.labkey.study.controllers.StudyController;
 import org.labkey.study.designer.StudyDesignInfo;
 import org.labkey.study.designer.StudyDesignManager;
@@ -368,22 +362,6 @@ public class StudyImpl extends ExtensibleStudyEntity<StudyImpl> implements Study
     public List<ParticipantCategoryImpl> getParticipantCategories(User user)
     {
         return ParticipantGroupManager.getInstance().getParticipantCategories(getContainer(), user);
-    }
-
-    public List<SpecimenRequestStatus> getSampleRequestStatuses(User user)
-    {
-        return SpecimenRequestManager.get().getRequestStatuses(getContainer(), user);
-    }
-
-    public Set<Integer> getSampleRequestStatusesInUse()
-    {
-        return SpecimenManager.getInstance().getRequestStatusIdsInUse(getContainer());
-    }
-
-    @Migrate // Use SettingsManager.get().getRepositorySettings(getContainer()) instead
-    public RepositorySettings getRepositorySettings()
-    {
-        return SettingsManager.get().getRepositorySettings(getContainer());
     }
 
     @Override
